+++ components/31-molecules/frequently-visited/frequently-visited.config.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = {
+  title: 'Frequently visited',
+  status: 'ready',
+  context: {
+    tags: ['Make an appointment', 'Opening hours', 'Events', 'Districts', 'Road works', 'Certificates', 'Job openings']
+  }
+};